Added mention of "agent-based modeling" to the masonclj entry.
Agent-based modeling and discrete-event simulation are overlapping
modeling traditions.  I would guess that anything you can do with a
library based in one tradition can be done with a library based in the
other tradition, but of course different libraries may emphasize
methods that are emphasized in a given tradition.  Some people are
familiar with one term or the other, so it's useful to mention both.  (I
wasn't always familiar with the term "discrete event simulation", even
though I was using MASON and its web page says it's a "discrete-event
multiagent simulation library"! :-)